feat: add EditProfileModal component, update user validation rules, and enhance picture handling

+++ src/components/Navbar/Navbar.tsx
@@ -1,15 +1,17 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import classes from './navbar.module.scss';
 import { Link } from 'react-router';
 import { useAuth } from '../../context/AuthContext';
-import { Button, Dropdown, Input, MenuProps, Modal, Typography, Spin, Divider, Form, Avatar, Upload } from 'antd';
+import { Avatar, Button, Divider, Dropdown, Form, Input, MenuProps, Modal, Typography, Upload } from 'antd';
 import { EditOutlined, LogoutOutlined, UploadOutlined, UserOutlined } from '@ant-design/icons';
 import { getPictureSrcUrl } from '@/utils/picture.utils.ts';
-import { AuthFormValidationRules } from '@/pages/AuthPage/auth.validation';
+import { UpdateUserValidationRules } from '@/pages/AuthPage/auth.validation';
 import { UsersService } from '@api/services/users.service.ts';
 import { AxiosError } from 'axios';
 import { UserUpdatePayload } from '@/models/user.model';
 import { FileService } from '@/api/services/file.service';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { QUERY_KEYS } from '@api/constants/query-keys.const.ts';
 // import { useQueryClient } from '@tanstack/react-query';
 // import { QUERY_KEYS } from '@/api/constants/query-keys.const';
 // import { AuthService } from '@/api/services/auth.service';
@@ -32,55 +34,57 @@ export interface RegistrationFormValues {
 }
 
 export const Navbar = () => {
-    const { loggedInUser, logout, handleAuthResponse } = useAuth();
+    const { loggedInUser, logout, doesUserHasGooglePicture, isGoogleUser } = useAuth();
 
-    const getDefaultUser = (): User | null => {
-        return loggedInUser ?? null;
-    };
-
-    const [user, setUser] = useState<User | null>(getDefaultUser());
     const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
     const [form] = Form.useForm();
-    const [editData, setEditData] = useState<User | null>(getDefaultUser());
 
     const [imageUrl, setImageUrl] = useState<string>('');
     const [selectedFile, setSelectedFile] = useState<File | null>(null);
     const [errorMessage, setErrorMessage] = useState<string | null>(null);
 
-    useEffect(() => {
-        if (loggedInUser) {
-            setUser(loggedInUser);
-            setEditData(loggedInUser);
-        }
-    }, [loggedInUser]);
-
     const showModal = () => setIsModalOpen(true);
     const handleCancel = () => setIsModalOpen(false);
+    const queryClient = useQueryClient();
+
+    const { mutateAsync } = useMutation({
+        mutationKey: [QUERY_KEYS.UPDATE_USER],
+        mutationFn: async (user: UserUpdatePayload) => {
+            return await UsersService.updateUser(loggedInUser!._id, user);
+        },
+        onSuccess: () => {
+            queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.LOGGED_IN_USER] });
+            handleCancel();
+        },
+    });
 
     const onFinish = async (values: RegistrationFormValues) => {
-        // const queryClient = useQueryClient();
-        let uploadedImageUrl;
-        let response;
-
-        if (selectedFile) {
-            const formData = new FormData();
-            formData.append('file', selectedFile);
-            const { data } = await FileService.handleUpload(formData);
-            uploadedImageUrl = data.url.split('public/')[1];
-        }
-
-        const user: UserUpdatePayload = {
-            username: values.username,
-            email: values.email,
-            picture: uploadedImageUrl,
-        };
+        let uploadedImageUrl: string | undefined = undefined;
 
         try {
             if (loggedInUser) {
-                await FileService.deleteFile(loggedInUser.picture);
-                const a = await UsersService.updateUser(loggedInUser._id, user);
-                // if (a) response = await AuthService.login(a!);
-                // if (response) handleAuthResponse(response);
+                const isPictureChanged = values.picture !== loggedInUser.picture;
+                const isEmailChanged = values.email !== loggedInUser.email;
+                const isUsernameChanged = values.username !== loggedInUser.username;
+
+                if (selectedFile) {
+                    const formData = new FormData();
+                    formData.append('file', selectedFile);
+                    const { data } = await FileService.handleUpload(formData);
+                    uploadedImageUrl = data.url.split('public/')[1];
+                }
+
+                const user: UserUpdatePayload = {
+                    username: isUsernameChanged ? values.username : undefined,
+                    email: isEmailChanged ? values.email : undefined,
+                    picture: uploadedImageUrl,
+                };
+
+                if (isPictureChanged && !doesUserHasGooglePicture && uploadedImageUrl) {
+                    await FileService.deleteFile(loggedInUser.picture);
+                }
+
+                await mutateAsync(user);
                 handleCancel();
             }
         } catch (e: unknown) {
@@ -90,20 +94,22 @@ export const Navbar = () => {
         }
     };
 
-    console.log(user ? getPictureSrcUrl(user.picture) : 'homo');
-
     const items: MenuProps['items'] = [
         {
             key: 'profile',
-            label: user ? (
+            label: loggedInUser ? (
                 <div className={classes.dropdownContainer}>
                     <div className={classes.profileImageContainer}>
-                        <img src={getPictureSrcUrl(user?.picture)} alt="User Avatar" className={classes.profileImage} />
+                        <img
+                            src={getPictureSrcUrl(loggedInUser?.picture)}
+                            alt="User Avatar"
+                            className={classes.profileImage}
+                        />
                     </div>
 
                     <div className={classes.textCenter}>
-                        <Text className={classes.profileName}>{user.username}</Text>
-                        <Text className={classes.profileEmail}>{user.email}</Text>
+                        <Text className={classes.profileName}>{loggedInUser.username}</Text>
+                        <Text className={classes.profileEmail}>{loggedInUser.email}</Text>
                     </div>
 
                     <Button
@@ -112,7 +118,7 @@ export const Navbar = () => {
                         className={classes.customizeProfileButton}
                         onClick={showModal}
                     >
-                        Customize Profile
+                        Edit User
                     </Button>
 
                     <Divider className={classes.divider} />
@@ -139,68 +145,78 @@ export const Navbar = () => {
             </Link>
 
             <div className={classes.right}>
-                {user ? (
-                    <Dropdown menu={{ items }} trigger={['click']} placement="bottomRight">
-                        <img src={getPictureSrcUrl(user?.picture)} alt="User Avatar" className={classes.avatarSmall} />
-                    </Dropdown>
-                ) : null}
-
-                <Modal title="Edit Profile" open={isModalOpen} onCancel={handleCancel} onOk={() => form.submit()}>
-                    <Form form={form} layout="vertical" onFinish={onFinish} className={classes.authForm}>
-                        <Form.Item name="picture">
-                            <Avatar
-                                src={imageUrl || undefined}
-                                icon={!imageUrl ? <UserOutlined /> : undefined}
-                                size={70}
+                {loggedInUser && (
+                    <>
+                        <Dropdown menu={{ items }} trigger={['click']} placement="bottomRight">
+                            <img
+                                src={getPictureSrcUrl(loggedInUser.picture)}
+                                alt="User Avatar"
+                                className={classes.avatarSmall}
                             />
-                        </Form.Item>
-                        <Form.Item
-                            name="upload"
-                            className={classes.uploadFormItem}
-                            rules={AuthFormValidationRules.picture}
-                        >
-                            <Upload
-                                maxCount={1}
-                                beforeUpload={(file) => {
-                                    setSelectedFile(file);
+                        </Dropdown>
 
-                                    const reader = new FileReader();
-                                    reader.onload = (e) => {
-                                        setImageUrl(e.target?.result as string);
-                                    };
-                                    reader.readAsDataURL(file);
-
-                                    return false;
-                                }}
-                                onRemove={() => {
-                                    setImageUrl('');
-                                    setSelectedFile(null);
-                                }}
-                                showUploadList={false}
+                        <Modal title="Edit User" open={isModalOpen} onCancel={handleCancel} onOk={() => form.submit()}>
+                            <Form
+                                form={form}
+                                layout="vertical"
+                                onFinish={onFinish}
+                                className={classes.authForm}
+                                initialValues={loggedInUser}
                             >
-                                <Button icon={<UploadOutlined />} block>
-                                    Upload Your Picture
-                                </Button>
-                            </Upload>
-                        </Form.Item>
-                        <Form.Item
-                            label="Username"
-                            name="username"
-                            className={classes.inputField}
-                            rules={AuthFormValidationRules.username}
-                        >
-                            <Input placeholder="Enter your username" />
-                        </Form.Item>
-                        <Form.Item
-                            label="Email"
-                            name="email"
-                            className={classes.inputField}
-                            rules={AuthFormValidationRules.email}
-                        >
-                            <Input placeholder="Enter your email" />
-                        </Form.Item>
-                    </Form>
-                </Modal>
+                                <Form.Item name="picture">
+                                    <Avatar
+                                        src={imageUrl || getPictureSrcUrl(loggedInUser.picture)}
+                                        icon={!imageUrl ? <UserOutlined /> : undefined}
+                                        size={70}
+                                    />
+                                </Form.Item>
+                                <Form.Item name="upload" className={classes.uploadFormItem}>
+                                    <Upload
+                                        maxCount={1}
+                                        beforeUpload={(file) => {
+                                            setSelectedFile(file);
+
+                                            const reader = new FileReader();
+                                            reader.onload = (e) => {
+                                                setImageUrl(e.target?.result as string);
+                                            };
+                                            reader.readAsDataURL(file);
+
+                                            return false;
+                                        }}
+                                        onRemove={() => {
+                                            setImageUrl('');
+                                            setSelectedFile(null);
+                                        }}
+                                        showUploadList={false}
+                                    >
+                                        <Button icon={<UploadOutlined />} block>
+                                            Upload Your Picture
+                                        </Button>
+                                    </Upload>
+                                </Form.Item>
+                                <Form.Item
+                                    label="Username"
+                                    name="username"
+                                    className={classes.inputField}
+                                    rules={UpdateUserValidationRules.username}
+                                >
+                                    <Input placeholder="Enter your username" />
+                                </Form.Item>
+                                {!isGoogleUser && (
+                                    <Form.Item
+                                        label="Email"
+                                        name="email"
+                                        className={classes.inputField}
+                                        rules={UpdateUserValidationRules.email}
+                                    >
+                                        <Input placeholder="Enter your email" />
+                                    </Form.Item>
+                                )}
+                            </Form>
+                        </Modal>
+                    </>
+                )}
             </div>
         </nav>
     );
